feat: remove firefighters from crew-selection if they are already assigned to the firetruck
